STORAGE: Shift mock server setup to another file
